// generated/src/app/api-models/commission-matrix-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CommissionMatrixDetail {
  matrixDef: string;
  matrixXSeq: number;
  matrixYSeq: number;
  matrixRate: number;
  matrixPct: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}